to_aur4.sh: support filter-branch of packages that were not valid
We run mksrcinfo for each filtered commit before pushing, and the AUR
expects all the commits to represent a valid package instance, but
some of the packages (e.g. ros-build-tools) were not valid at some
point. To fix this, we automatically squash the first commits until
the first valid one.